+++ lib/cye/include/cye/instance.hpp
@@ -26,9 +26,9 @@ class Instance {
   [[nodiscard]] inline auto cargo_capacity() const { return cargo_capacity_; }
   [[nodiscard]] inline auto battery_capacity() const { return battery_capacity_; }
   [[nodiscard]] inline auto closest_charging_station(size_t node_id) const -> size_t { return closest_charging_station_[node_id]; }
-  [[nodiscard]] inline auto distance(size_t node1_id, size_t node2_id) const -> float {
+  [[nodiscard]] inline auto distance(size_t node1_id, size_t node2_id) const -> double {
     if (node1_id == node2_id) [[unlikely]] {
-      return 0.f;
+      return 0.0;
     }
     if (node1_id > node2_id) std::swap(node1_id, node2_id);
 
@@ -52,27 +52,27 @@ class Instance {
   auto update_closest_charging_station_() -> void;
 
   std::string name_;
-  float optimal_value_;
+  double optimal_value_;
   size_t minimum_route_cnt_;
-  float cargo_capacity_;
-  float battery_capacity_;
-  float energy_consumption_;
+  double cargo_capacity_;
+  double battery_capacity_;
+  double energy_consumption_;
   size_t customer_cnt_;
   size_t charging_station_cnt_;
 
   std::vector<Node> nodes_;
-  std::vector<float> distance_cache_;
+  std::vector<double> distance_cache_;
   std::vector<size_t> closest_charging_station_;
 };
 
 template <serial::Value V>
 Instance::Instance(V &&value)
     : name_(value["name"].template get<std::string_view>()),
-      optimal_value_(value["optimalValue"].template get<float>()),
+      optimal_value_(value["optimalValue"].template get<double>()),
       minimum_route_cnt_(value["minimumRouteCnt"].template get<size_t>()),
-      cargo_capacity_(value["cargoCapacity"].template get<float>()),
-      battery_capacity_(value["batteryCapacity"].template get<float>()),
-      energy_consumption_(value["energyConsumption"].template get<float>()),
+      cargo_capacity_(value["cargoCapacity"].template get<double>()),
+      battery_capacity_(value["batteryCapacity"].template get<double>()),
+      energy_consumption_(value["energyConsumption"].template get<double>()),
       customer_cnt_(value["customerCnt"].template get<size_t>()),
       charging_station_cnt_(value["chargingStationCnt"].template get<size_t>()),
       nodes_(value["nodes"].template get<std::vector<Node>>()) {
